[2/2] OmniGears: lock screen config
configure time and date display on lock screen
-select what to show time/date/alarm/nothing
-configure color/font/shadow
-add shortcut hide
-show statusbar time if lock screen time is hidden

add FontPreference and NumberPickerPreference
to pick a font - value stored is the path to the ttf file use
Typeface.createFromFile(file) to apply to a text

And while we are here fix the shadow slider settings

+++ src/org/omnirom/omnigears/interfacesettings/BarsSettings.java
@@ -126,7 +126,7 @@ public class BarsSettings extends SettingsPreferenceFragment implements
         mHeaderShadow = (SeekBarPreference) findPreference(CUSTOM_HEADER_IMAGE_SHADOW);
         final int headerShadow = Settings.System.getInt(getContentResolver(),
                 Settings.System.STATUS_BAR_CUSTOM_HEADER_SHADOW, 0);
-        mHeaderShadow.setValue((int)((headerShadow / 255) * 100));
+        mHeaderShadow.setValue((int)(((double) headerShadow / 255) * 100));
         mHeaderShadow.setOnPreferenceChangeListener(this);
     }
